requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
openai==1.68.2
chromadb==0.6.3
PyPDF==5.4.0
python-dotenv==1.0.1
